frontend/requirements.txt
gradio[mcp]
requests
websocket-client
imageio
imageio-ffmpeg
PyYAML
pydantic
pydub
pyperclip
hf_xet
tqdm
pymediainfo
aria2